+++ sample-apps/webhooks-app/src/js/contacts-controller.js
@@ -2,6 +2,7 @@ const _ = require('lodash')
 const express = require('express')
 const router = new express.Router()
 const dbHelper = require('./db-helper')
+const utils = require('./utils')
 
 const EVENTS_COUNT_PER_PAGE = 25
 
@@ -81,9 +82,9 @@ exports.getRouter = () => {
             console.log('Calling hubspotClient.crm.contacts.batchApi.read API method. Retrieve contacts.')
             // Get a batch of contacts by id's
             // POST /crm/v3/objects/contacts/batch/read
-            // https://developers.hubspot.com/docs-beta/crm/contacts
+            // https://developers.hubspot.com/docs/api/crm/contacts
             const contactsResponse = await req.hubspotClient.crm.contacts.batchApi.read(false, batchRead)
-            console.log(JSON.stringify(contactsResponse, null, 2))
+            utils.logJson(contactsResponse.body)
 
             const events = await dbHelper.getEvents(contactIds)
             const contacts = prepareContactsForView(events, contactsResponse.body.results)