+++ CHANGELOG.md
@@ -42,6 +42,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   24. getSubscription => getById (webhooks.subscriptionsApi)
   25. getSubscriptions => getAll (webhooks.subscriptionsApi)
   26. updateSubscription => update (webhooks.subscriptionsApi)
+  
+## [1.1.0-beta] - 2020-04-24
+  
+### Added
+  
+  - New updateBatch method to webhooks.subscriptionsApi
+  - updated timelineModels
+
+### Updated
+
+  - extended webhooks-sample-app to setup application webhooks configuration after successful authorization
+  
+### Fixed
+
+  - 'apiKey' authorization for webhooks clients  
 
 [unreleased]: https://github.com/HubSpot/hubspot-api-nodejs/compare/v1.0.0-beta...HEAD
 [1.0.0-beta]: https://github.com/HubSpot/hubspot-api-nodejs/releases/tag/v1.0.0-beta
+[1.1.0-beta]: https://github.com/HubSpot/hubspot-api-nodejs/releases/tag/v1.1.0-beta